refactor: 상속 단계 추가해 중복 코드 제거
