+++ src/main/java/domain/piece/Cannon.java
@@ -1,15 +1,12 @@
 package domain.piece;
 
 import domain.Country;
-import domain.Direction;
 import domain.Position;
 import domain.state.State;
 import java.util.List;
 import java.util.Map;
 
-public class Cannon extends Piece {
-    private static final String ONLY_MOVE_STRAIGHT = "[ERROR] 포는 직선으로만 이동 가능합니다.";
-    private static final String FIXED_DIRECTION = "[ERROR] 포는 하나의 방향으로만 이동 가능합니다.";
+public class Cannon extends MoveStraightPiece {
     private static final String INVALID_JUMP_PIECE_COUNT = "[ERROR] 포는 하나의 기물만 뛰어 넘을 수 있습니다.";
     private static final String CANNOT_JUMP_CANNON = "[ERROR] 포는 포를 뛰어 넘을 수 없습니다.";
     private static final String CANNOT_KILL_CANNON = "[ERROR] 포는 포를 잡을 수 없습니다.";
@@ -19,20 +16,6 @@ public class Cannon extends Piece {
 
     public Cannon(Country country) {
         super(new PieceInfo(PieceType.CANNON, country));
-    }
-
-    @Override
-    public void validateDirections(List<Direction> directions) {
-        Direction oneSide = directions.getFirst();
-        boolean allSameDirection = directions.stream()
-                .allMatch(direction -> direction.equals(oneSide));
-        if (!allSameDirection) {
-            throw new IllegalArgumentException(FIXED_DIRECTION);
-        }
-        // 궁성 영역 생각하지 않음
-        if (oneSide.isDiagonal()) {
-            throw new IllegalArgumentException(ONLY_MOVE_STRAIGHT);
-        }
     }
 
     @Override
