+++ src/test/java/domain/piece/GeneralTest.java
@@ -82,7 +82,7 @@ public class GeneralTest {
 
         assertThatThrownBy(() -> general.path(from, to))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 궁은 한 칸만 이동할 수 있습니다.");
+                .hasMessage("[ERROR] 해당 기물은 한 칸만 이동할 수 있습니다.");
     }
 
     @Test
@@ -95,7 +95,7 @@ public class GeneralTest {
 
         assertThatThrownBy(() -> general.path(from, to))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 궁은 직선으로만 이동 가능합니다.");
+                .hasMessage("[ERROR] 해당 기물은 직선으로만 이동 가능합니다.");
     }
 
     @Test
